Implement unit test of MockItemRepository and add test dependencies

+++ app/src/main/java/com/example/pixbayphoto/domain/model/Item.kt
@@ -20,7 +20,7 @@ data class Item(
     val collections: Int,
     val likes: Int,
     val comments: Int,
-    val userId: Int,
+    val userId: Long,
     val user: String,
     val userImageUrl: String,
     val noAiTraining: Boolean,
